+++ lib/rpmsg_lite/porting/environment/rpmsg_env_freertos.c
@@ -77,7 +77,7 @@ StaticEventGroup_t event_group_static_context;
 #define RL_ENV_MAX_MUTEX_COUNT (10)
 
 /* Max supported ISR counts */
-#define ISR_COUNT (64U)
+#define ISR_COUNT RL_PLATFORM_MAX_ISR_COUNT
 /*!
  * Structure to keep track of registered ISR's.
  */